app.js renamed to index.js
*When updating / deleting / creating a user, nothing is returned
*The connection to the database has been moved to index.js
