+++ package.json
@@ -1,7 +1,7 @@
 {
     "scripts": {
-        "start": "node app.js",
-        "dev": "nodemon app.js"
+        "start": "node index.js",
+        "dev": "nodemon index.js"
     },
     "dependencies": {
         "body-parser": "^1.19.0",
